# document/openapi.yaml
openapi: 3.1.0
info:
  title: Cash Flow Simulator API
  version: 0.1.0
  summary: MVP API specification for the cash flow simulation system
  description: |
    Frontend handoff specification for the current MVP API.

    Maintenance rule:
    - When request/response shapes, query parameters, status codes, or endpoint behavior change,
      update this OpenAPI document in the same change set.
    - Treat this file as the canonical contract for frontend implementation and API review.
servers:
  - url: http://127.0.0.1:3000
    description: Local development
...
paths:
...
  /api/credit-cards/{id}:
    patch:
      tags: [CreditCards]
      summary: Update credit card
      operationId: updateCreditCard
      parameters:
        - name: id
          in: path
          required: true
          schema:
            $ref: "#/components/schemas/UuidLike"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreditCardUpdateRequest"
      responses:
        "200":
          description: Updated credit card
          content:
            application/json:
              schema:
                type: object
                required: [creditCard]
                properties:
                  creditCard:
                    $ref: "#/components/schemas/CreditCard"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      tags: [CreditCards]
      summary: Delete credit card
      operationId: deleteCreditCard
      parameters:
        - name: id
          in: path
          required: true
          schema:
            $ref: "#/components/schemas/UuidLike"
      responses:
        "200":
          description: Deleted credit card
          content:
            application/json:
              schema:
                type: object
                required: [creditCard]
                properties:
                  creditCard:
                    $ref: "#/components/schemas/CreditCard"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/transactions:
    get:
      tags: [Transactions]
      summary: List transactions
      operationId: listTransactions
      parameters:
        - $ref: "#/components/parameters/StartDate"
        - $ref: "#/components/parameters/EndDate"
      responses:
        "200":
          description: Transaction list
          content:
            application/json:
              schema:
                type: object
                required: [transactions]
                properties:
                  transactions:
                    type: array
                    items:
                      $ref: "#/components/schemas/Transaction"
    post:
      tags: [Transactions]
      summary: Create transaction
      operationId: createTransaction
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TransactionCreateRequest"
      responses:
        "201":
          description: Created transaction
          content:
            application/json:
              schema:
                type: object
                required: [transaction]
                properties:
                  transaction:
                    $ref: "#/components/schemas/Transaction"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /api/scheduled-events:
    get:
      tags: [ScheduledEvents]
      summary: List scheduled events
      operationId: listScheduledEvents
      parameters:
        - $ref: "#/components/parameters/StartDate"
        - $ref: "#/components/parameters/EndDate"
        - name: isActive
          in: query
          required: false
          description: Filter by active flag
          schema:
            type: boolean
      responses:
        "200":
          description: Scheduled event list
          content:
            application/json:
              schema:
                type: object
                required: [scheduledEvents]
                properties:
                  scheduledEvents:
                    type: array
                    items:
                      $ref: "#/components/schemas/ScheduledEvent"
    post:
      tags: [ScheduledEvents]
      summary: Create scheduled event
      operationId: createScheduledEvent
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ScheduledEventCreateRequest"
      responses:
        "201":
          description: Created scheduled event
          content:
            application/json:
              schema:
                type: object
                required: [scheduledEvent]
                properties:
                  scheduledEvent:
                    $ref: "#/components/schemas/ScheduledEvent"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  parameters:
    StartDate:
      name: startDate
      in: query
      required: false
      description: Inclusive start date filter (`YYYY-MM-DD`)
      schema:
        type: string
        format: date
    EndDate:
      name: endDate
      in: query
      required: false
      description: Inclusive end date filter (`YYYY-MM-DD`)
      schema:
        type: string
        format: date
...
  schemas:
    ErrorResponse:
      type: object
      required: [message]
      properties:
        message:
          type: string
...
    JsonTags:
      type: object
      additionalProperties: true
      description: Flexible tag payload. Current usage commonly includes `category` and `project`.
...
    CreditCard:
      type: object
      required:
        [id, name, closingDay, paymentDay, settlementAccountId, currency, isDefault, createdAt, updatedAt]
      properties:
        id:
          $ref: "#/components/schemas/UuidLike"
        name:
          type: string
        closingDay:
          type: integer
          minimum: 1
          maximum: 31
        paymentDay:
          type: integer
          minimum: 1
          maximum: 31
        settlementAccountId:
          anyOf:
            - $ref: "#/components/schemas/UuidLike"
            - type: "null"
        currency:
          type: string
          minLength: 3
          maxLength: 3
        isDefault:
          type: boolean
        createdAt:
          type: string
        updatedAt:
          type: string
...
    CreditCardUpdateRequest:
      type: object
      properties:
        name:
          type: string
        closingDay:
          type: integer
          minimum: 1
          maximum: 31
        paymentDay:
          type: integer
          minimum: 1
          maximum: 31
        settlementAccountId:
          anyOf:
            - $ref: "#/components/schemas/UuidLike"
            - type: "null"
        currency:
          type: string
          minLength: 3
          maxLength: 3
        isDefault:
          type: boolean
    Transaction:
      type: object
      required:
        [id, date, amount, accountId, tags, cardId, memo, orderIndex, createdAt, updatedAt]
      properties:
        id:
          $ref: "#/components/schemas/UuidLike"
        date:
          type: string
          format: date
        amount:
          type: string
          description: Signed decimal string
        accountId:
          anyOf:
            - $ref: "#/components/schemas/UuidLike"
            - type: "null"
        tags:
          $ref: "#/components/schemas/JsonTags"
        cardId:
          anyOf:
            - $ref: "#/components/schemas/UuidLike"
            - type: "null"
        memo:
          type: string
        orderIndex:
          type: integer
          minimum: 0
        createdAt:
          type: string
        updatedAt:
          type: string
    TransactionCreateRequest:
      type: object
      required: [date, amount]
      properties:
        date:
          type: string
          format: date
        amount:
          oneOf:
            - type: string
            - type: number
        accountId:
          anyOf:
            - $ref: "#/components/schemas/UuidLike"
            - type: "null"
        tags:
          $ref: "#/components/schemas/JsonTags"
        cardId:
          anyOf:
            - $ref: "#/components/schemas/UuidLike"
            - type: "null"
        memo:
          type: string
          default: ""
        orderIndex:
          type: integer
          minimum: 0
          default: 0
    ScheduledEvent:
      type: object
      required:
        [id, name, startDate, endDate, recurrenceRule, amount, accountId, tags, cardId, isActive, orderIndex, createdAt, updatedAt]
      properties:
        id:
          $ref: "#/components/schemas/UuidLike"
        name:
          type: string
        startDate:
          type: string
          format: date
        endDate:
          anyOf:
            - type: string
              format: date
            - type: "null"
        recurrenceRule:
          anyOf:
            - type: string
            - type: "null"
        amount:
          type: string
        accountId:
          anyOf:
            - $ref: "#/components/schemas/UuidLike"
            - type: "null"
        tags:
          $ref: "#/components/schemas/JsonTags"
        cardId:
          anyOf:
            - $ref: "#/components/schemas/UuidLike"
            - type: "null"
        isActive:
          type: boolean
        orderIndex:
          type: integer
          minimum: 0
        createdAt:
          type: string
        updatedAt:
          type: string
    ScheduledEventCreateRequest:
      type: object
      required: [name, startDate, amount]
      properties:
        name:
          type: string
        startDate:
          type: string
          format: date
        endDate:
          anyOf:
            - type: string
              format: date
            - type: "null"
        recurrenceRule:
          anyOf:
            - type: string
            - type: "null"
        amount:
          oneOf:
            - type: string
            - type: number
        accountId:
          anyOf:
            - $ref: "#/components/schemas/UuidLike"
            - type: "null"
        tags:
          $ref: "#/components/schemas/JsonTags"
        cardId:
          anyOf:
            - $ref: "#/components/schemas/UuidLike"
            - type: "null"
        isActive:
          type: boolean
          default: true
        orderIndex:
          type: integer
          minimum: 0
          default: 0
...
    UuidLike:
      type: string
      description: UUID-like identifier currently accepted by the backend validation layer
      pattern: "^[0-9a-fA-F-]{36}$"